Revert "feat: Add utils and lang_service"
This reverts commit f1ac5eae0ed1d9420748224913f2bb5269ccadb5.
